Clave de alumno

+++ src/main/java/org/geducativo/geradmin3/RepositorioClave.java
@@ -5,5 +5,5 @@ import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.data.repository.Repository;
 
-public interface RepositorioClave extends MongoRepository<Clave, Long> {
+public interface RepositorioClave extends MongoRepository<Clave, Integer> {
 }
